#161 [feat] : move to top usecase 분리
